+++ src/lib/oauth/providers/index.js
@@ -24,6 +24,7 @@ import clinepass from "./clinepass.js";
 import gitlab from "./gitlab.js";
 import codebuddyCn from "./codebuddy-cn.js";
 import codebuddyIntl from "./codebuddy-intl.js";
+import workbuddy from "./workbuddy.js";
 import kimchi from "./kimchi.js";
 import trae from "./trae.js";
 import windsurf from "./windsurf.js";
@@ -51,7 +52,7 @@ const PROVIDERS = {
   gitlab,
   "codebuddy-cn": codebuddyCn,
   "codebuddy-intl": codebuddyIntl,
-  kimchi,
+  workbuddy,
   trae,
   windsurf,
   zed,